Add volume_mesh to BoundaryData
Also change volume_mesh_ghost_cell_data to std::optional.
Now the item for Tags::NeighborMesh is always filled with the
communicated volume_mesh which is the active mesh of the neighbor.
The item for Tags::MeshForGhostData is set to the communicated
volume_mesh_ghost_cell_data, and is only used when doing DG-subcell.