Fix printing tensors in GDB on mbot
There were a number of subtleties, and unfortunately the GDB documentation is
essentially non-existent, but this resolves the issue. It does make the output
not as pretty since I need to print the storage index instead of the component
suffix. I don't know of an easy way around this, though.