Fix yaml issues
Fix loading and writing of multiple yaml documents
in the same stream. Previously, the existing tutorial
presumably worked with a single document yaml stream.
The current `BinaryBlackHole.yaml` has multiple docs.
This change allows such streams to be read and edited
in this example notebook.

Make minor changes in path names to make it more informative.