Always redo step if size change is desired
The reason for accepting steps and changing the next is to avoid
discarding the work involved in taking steps, but we only change the
size at the start of the step, so the only work that has been done is
a single update of the variables.  Taking a larger step to avoid an
extra RHS evaluation is better.

Only allow the size to change once per step to avoid potential
infinite loops.