Exporter: add mode to extrapolate from nearest element
Uses the Lagrange basis in the nearest element to extrapolate into the
excision region. This can be unstable when we want to extrapolate
far into the excision because we're extrapolating over many logical element sizes.
However, it can be useful for extrapolating only very slightly into the
excision, e.g. if the excision in the initial data is the apparent horizon
and the evolution needs a little space around it to find and track the
horizon.