Error if we restarted with overlay and can't find the next phase
Before this commit, if you restarted from a checkpoint and overlayed the
PhaseChangeAndTriggers removing the phase change that caused the checkpoint and
restart then the simulation would restart and error that the phase
UpdateOptionsAtRestartFromCheckpoint is not in the default phase order (which it
shouldn't be!). This commit adds a check and error that clarifies what went
wrong.