Replace LinearLeastSquares with free functions
The class was supposed to cache allocations between calls for
efficiency, but it still did a separate workspace allocation for every
point in vector mode and the class was only ever used once before
being discarded.  Furthermore, the memory management was implemented
incorrectly and the code would crash if an instance was moved.