Generalize Neighbors for non-conforming Blocks
With non-conforming Blocks, a Block may have multiple neighbors
in a direction, and those neighbors need not share a common
Orientation.  Thus store an OrientationMap indexed by the id of
the Block.  This is sufficient for Elements as well since all
Elements in a given Block are aligned with each other.

Note this requires a block id to be passed to fetch the
correct OrientationMap from the Neighbors.

Also, for convenience, store a boolean on whether or not the
neighbors are conforming to the host.

Also add sanity checks for consistent member variables.
Check that conforming neighbors must be in the same Block.
Check that all ids have an orientation.
This required fixing several tests that violated these
assumptions.  (The tests were not wrong, they just constructed
multiple neighbors from different Blocks.)