Add states DeltaRDriftInward, DeltaRNoDrift
These correspond to State 3 and State 4 in
SpEC.  These states can prevent the excision
boundary from getting too close to the horizon.

NO TESTS are added in this commit; see the next commit.

The current states (Initial,DeltaRDriftOutward,AhSpeed,DeltaR) have
NOT been updated to transition to the new states in this commit.

This commit involves:
 - Adding the two new states
 - Updating Info to hold things that the new states use.
 - Updating State and StateHistory to know about the new states.