Add struct InterpolatedBoundaryData
InterpolatedBoundaryData will be used to communicate boundary data
that is communicated from multiple non-conforming neighbors to a
target Element.  Each of the multiple neighbors will interpolate
to a subset of the points of boundary mortar of the target.  This
requires sending the interpolated data as well as the offsets into
a DataVector covering the target Mesh.  The target Mesh is also in
case the target Element undergoes p-refinement, in which case the
target can no longer use the interpolated data.