Change orientation of NeighborMesh
For the elliptic code, neither the host nor neighbor frame seems
particularly more natural, but for evolution code the host frame is
easier to work with, so switch to that.

The evolution code currently using the neighbor mesh doesn't handle
orientations right for either convention, so it isn't touched here and
will be fixed in subsequent commits.

NewNeighborIds.cpp has large reformatting because the actually
significant small change caused unused-variable warnings in gcc.