BlockLogicalCoordinates: set initial block order
If an empty block order is provided, then initially fill it
with the list of blocks in the domain.
This avoids having to fill it with std::iota initially, which
requires knowledge of the number of blocks in the domain,
which in turn requires opening volume data files if
interpolating from disk. So this just makes it easier.