Add tests for NP spin coeffs
- AnalyticTestVolumeNewmanPenroseSpinCoefficients

We use the analytical values of the Newman-Penrose spin coefficients in the
Schwarzschild background as a simple test of the spin coefficient code.  Almost
all of them vanish, but they may have nontrivial cancellations.  Three
coefficients are nonvanishing, and are checked against their analytical
expression in terms of 1/r, derived from OneMinusY.

- Add NP spin coefficients to Test_Tags

- Add test_compute_tag's for NP spin coeffs in Test_NewmanPenrose

- Add python regressions tests for NP spin coeffs