Fix variable-order LTS assertion failure
The TimeStepId from the MortarNextTemporalId and the newly received id
represent the same temporal event, but the one from
MortarNextTemporalId may have an out-of-date slab if we are at a slab
boundary and the slab size was changed (including roundoff-level
changes related to the slab size goal).  Having the wrong slab would
cause an assertion failure when trying to subtract times in the Adams
integrators, but, because of step alignment requirements and some
arbitrary choices made in the integrators, this value can only be used
when the remote side is running at higher-order than the local side.