Add enum domain::FaceType and use it in Element
The FaceType labels whether the face of an Element abuts the
external boundary, neighboring Elements, or neither of those.
The latter occurs for topologies other than I1 such as the
angular faces of a spherical shell.  For neighboring Elements,
the FaceType will determine how boundary information is exchanged
between the Elements.

The Element now stores a DirectionMap of FaceType for convenience.