Optimizations in TensorYlmFilter.
Add "if" statements so as to avoid computing matrix elements
that are zero at an early stage.  This involves computing a few
threejs and other quantities earlier.

Also, remove "if" statements if they can be handled by changing
the bounds of loops.

And also for things like `auto a = b * c * d * e * f`; where all
the factors are doubles except one of them is complex, put the
complex one at the end to avoid unnecessary complex arithmetic.