Stop serializing FoT update backlogs
FoT serialization is done when writing volume data, so must be safe to
run in parallel with FoT updates.  The backlog structures are not safe
to access in this manner.

Omitting the backlogs is fine, since they should always be empty when
checkpointing, as they are just for handling out-of-order messages,
and the things that read FoTs from the volume file can't make use of
the information.