Move CCE objects to the global cache
For consistency with other systems, which will help with code sharing.

I think the time stepper used in InitializeWorldtubeBoundary is
intended to be the one used in the CCE evolution, so it is unclear to
me why it was parsed as GTS.  I have changed it to actually use the
CCE object.