Fix quadrature weights for Chebyshev
The weights should not be computed from the inverse weight function
and the integration weights as the weight function cannot be
represented in the basis.  Instead they can be computed from
multiplying the nodal_to_modal_matrix by a row vector.

For all the other current bases, the integration and quadrature
weights are the same.