Record volume observation FoT data on elements
When we clean old FoT data, it will be easy to make sure the elements
are done with it, but harder to keep track on other components.

The global FoT data will be fairly useless if cleaning is enabled, but
it can continue working as it currently is when there is no cleaning.

Also add a few missing tag tests.