Keep slightly more data in ThreadsafeList trunc
Requesting the end of an interval still accesses the interval before
it, even though the value doesn't matter.  We could do extra checks to
avoid the access, but adding more checks in the common case to avoid
keeping one extra interval when truncating exactly at a boundary isn't
worth it.