Improve remove_duplicates implementation
Using list_contains is faster than using none.

For GCC 15, this is about a 10x speedup.  Clang 20 seems to have worse
than quadratic scaling with the new method, but the constant is
improved enough that it is still better than the old algorithm for
lengths that don't hit the recursion limit.  Both compilers showed
comparable performance with quadratic scaling with the old
implementation.